Added company

+++ app/services/users.py
@@ -5,7 +5,7 @@ from fastapi import HTTPException
 from passlib.context import CryptContext
 from starlette import status
 
-from app.conf.messages import ERROR_USER_NOT_FOUND, ERROR_ACCOUNT_EXISTS
+from app.conf.messages import ERROR_USER_NOT_FOUND, ERROR_ACCOUNT_EXISTS, ERROR_ACCESS
 from app.schemas.user_schemas import SignUpRequestModel, UserUpdate
 from app.utils.repository import AbstractRepository
 
@@ -33,19 +33,23 @@ class UsersService:
             raise HTTPException(status_code=404, detail=ERROR_USER_NOT_FOUND)
         return await self.users_repo.find_by_filter(filter_by)
 
-    async def update_user(self, user_id: int, user_data: UserUpdate):
+    async def update_user(self, user_id: int, user_data: UserUpdate, current_user: int):
         user_dict = user_data.model_dump()
         filter_by = {"id": user_id}
         user = await self.users_repo.find_by_filter(filter_by)
         if user is None:
             raise HTTPException(status_code=404, detail=ERROR_USER_NOT_FOUND)
+        if current_user.id != user_id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=ERROR_ACCESS)
         return await self.users_repo.update_by_filter(filter_by, user_dict)
 
-    async def delete_user(self, user_id: int):
+    async def delete_user(self, user_id: int, current_user: int):
         filter_by = {"id": user_id}
         user = await self.users_repo.find_by_filter(filter_by)
         if user is None:
             raise HTTPException(status_code=404, detail=ERROR_USER_NOT_FOUND)
+        if current_user.id != user_id:
+            raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail=ERROR_ACCESS)
         await self.users_repo.delete_by_id(user_id)
 
     async def find_user_by_email(self, email: str):
